requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.3
django-celery-beat==2.4.0
django-environ==0.11.2
django-timezone-field==5.0
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
kombu==5.2.4
outcome==1.2.0
packaging==21.3
prompt-toolkit==3.0.33
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.5
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wsproto==1.2.0
